% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_dir.R
\name{capsule_dir}
\alias{capsule_dir}
\title{Get or create the rtimecapsule directory}
\usage{
capsule_dir()
}
\description{
Internal helper to locate the \code{.rtimecapsule} directory
inside the current project.
}
\keyword{internal}
