% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{file_hash}
\alias{file_hash}
\title{Compute file hash}
\usage{
file_hash(file, root = project_root())
}
\arguments{
\item{file}{File path relative to project root}

\item{root}{Project root path. Defaults to current RStudio project.}
}
\value{
Character string of the file's MD5 hash, or NA if the file does not exist
}
\description{
Compute file hash
}
\examples{
\dontrun{
file_hash("script.R")
}
}
\keyword{internal}
