% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_backups.R
\name{list_backups}
\alias{list_backups}
\title{List backups for a file}
\usage{
list_backups(file, root = project_root())
}
\arguments{
\item{file}{File path relative to project root}

\item{root}{Project root path. Defaults to current RStudio project.}
}
\value{
Character vector of backup filenames
}
\description{
List backups for a file
}
\keyword{internal}
