% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_root}
\alias{project_root}
\title{Detect the current project root}
\usage{
project_root()
}
\value{
Absolute path to project root
}
\description{
Walks up the directory tree until an \code{.Rproj} file is found.
Prevents execution inside an R package source directory.
}
\keyword{internal}
