% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore_file}
\alias{restore_file}
\title{Restore a file from backup}
\usage{
restore_file(file)
}
\arguments{
\item{file}{File path relative to project root}
}
\value{
Invisibly TRUE
}
\description{
Restore a file from backup
}
