% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_last_crash.R
\name{restore_last_crash}
\alias{restore_last_crash}
\title{Restore all files from last backup}
\usage{
restore_last_crash()
}
\value{
Invisibly TRUE
}
\description{
Restore all files from last backup
}
