% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start_autobackup}
\alias{start_autobackup}
\title{Start rtimecapsule auto-backup (non-blocking)}
\usage{
start_autobackup(interval = 2, files = NULL)
}
\arguments{
\item{interval}{Seconds between checks}

\item{files}{NULL (default), a file extension (e.g. ".R"),
or a character vector of file names}
}
\value{
Invisibly TRUE
}
\description{
Runs a background task that mirrors selected files
into the .rtimecapsule folder whenever they change.
}
