% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{tracked_files}
\alias{tracked_files}
\title{List tracked files in the project}
\usage{
tracked_files(root = project_root(), files = NULL)
}
\arguments{
\item{root}{Project root path}

\item{files}{NULL (default), a single extension like ".R",
or a character vector of specific file names}
}
\value{
Character vector of relative file paths
}
\description{
Determines which files should be tracked based on user input.
}
\examples{
\dontrun{
# Only works inside an RStudio project
tracked_files()
}
}
\keyword{internal}
